Separate Auth functions from Account function #4
